+++ src/main/java/com/lffblk/restcup/model/entity/User.java
@@ -12,12 +12,12 @@ public class User {
     private String firstName;
     private String lastName;
     private String gender;
-    private long birthDate;
+    private int birthDate;
     private String email;
 
     public User() {}
 
-    public User(int id, String firstName, String lastName, String gender, long birthDate, String email) {
+    public User(int id, String firstName, String lastName, String gender, int birthDate, String email) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -58,11 +58,11 @@ public class User {
         this.gender = gender;
     }
 
-    public long getBirthDate() {
+    public int getBirthDate() {
         return birthDate;
     }
 
-    public void setBirthDate(long birthDate) {
+    public void setBirthDate(int birthDate) {
         this.birthDate = birthDate;
     }
